+++ ScriptLogger/Functions/Write-WarningLog.ps1
@@ -35,7 +35,7 @@ function Write-WarningLog
         $Name = 'Default',
 
         # The warning message.
-        [Parameter(Mandatory=$true, Position = 0, ValueFromPipeline = $true)]
+        [Parameter(Mandatory = $true, Position = 0, ValueFromPipeline = $true)]
         [System.String[]]
         $Message
     )